rename module hdfreadcontrol.py -> hdfreadcontrols.py and file bapsflib._hdf.utils.hdfreadcontrol.rst -> bapsflib._hdf.utils.hdfreadcontrols.rst
