+++ bapsflib/_hdf/utils/helpers.py
@@ -36,7 +36,7 @@ def build_shotnum_dset_relation(
     Compares the **shotnum** numpy array to the specified dataset,
     **dset**, to determine which indices contain the desired shot
     number(s)
-    [for :class:`~bapsflib._hdf.utils.hdfreadcontrol.HDFReadControls`].
+    [for :class:`~bapsflib._hdf.utils.hdfreadcontrols.HDFReadControls`].
     As a results, two numpy arrays are returned which satisfy the rule::
 
         shotnum[sni] = dset[index, shotnumkey]
@@ -357,7 +357,7 @@ def condition_controls(hdf_file: File,
                        controls: Any) -> List[Tuple[str, Any]]:
     """
     Conditions the **controls** argument for
-    :class:`~.hdfreadcontrol.HDFReadControls` and
+    :class:`~.hdfreadcontrols.HDFReadControls` and
     :class:`~.hdfreaddata.HDFReadData`.
 
     :param hdf_file: HDF5 object instance
@@ -478,7 +478,7 @@ def condition_shotnum(shotnum: Any,
                       shotnumkey_dict: Dict[str, str]) -> np.ndarray:
     """
     Conditions the **shotnum** argument for
-    :class:`~bapsflib._hdf.utils.hdfreadcontrol.HDFReadControls` and
+    :class:`~bapsflib._hdf.utils.hdfreadcontrols.HDFReadControls` and
     :class:`~bapsflib._hdf.utils.hdfreaddata.HDFReadData`.
 
     :param shotnum: desired HDF5 shot numbers
