+++ bapsflib/_hdf/utils/tests/test_hdfreadcontrol.py
@@ -22,7 +22,7 @@ from unittest import mock
 
 from . import (TestBase, with_bf)
 from ..file import File
-from ..hdfreadcontrol import HDFReadControls
+from ..hdfreadcontrols import HDFReadControls
 
 
 class TestHDFReadControl(TestBase):
