--- docs/src/bapsflib._hdf.utils.hdfreadcontrol.rst
@@ -1,11 +1,0 @@
-bapsflib\.\_hdf\.utils\.hdfreadcontrol
-======================================
-
-.. automodule:: bapsflib._hdf.utils.hdfreadcontrol
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-    .. rubric:: Classes
-
-    .. autosummary:: HDFReadControls
