+++ docs/src/bapsflib._hdf.utils.hdfreadcontrols.rst
@@ -1,0 +1,11 @@
+bapsflib\.\_hdf\.utils\.hdfreadcontrols
+=======================================
+
+.. automodule:: bapsflib._hdf.utils.hdfreadcontrols
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+    .. rubric:: Classes
+
+    .. autosummary:: HDFReadControls
