+++ docs/using_lapd/read_from_digi.inc.rst
@@ -254,7 +254,7 @@ Adding Control Device Data
 Adding control device data to a digitizer dataset is done with the
 keyword :data:`add_controls`.  Specifying :data:`add_controls` will
 trigger a call to the
-:class:`~bapsflib._hdf.utils.hdfreadcontrol.HDFReadControls` class and
+:class:`~bapsflib._hdf.utils.hdfreadcontrols.HDFReadControls` class and
 extract the desired control device data.
 :class:`~bapsflib._hdf.utils.hdfreaddata.HDFReadData` then compares and
 mates that control device data with the digitizer data according to the
